java GUI chess
This is a Java GUI to write a China chess project based on only set up a framework for many features not perfect, but the design knowledge is already here. This code, and not what technical content, personally feel that there are many places need to be improved. The main method to run inside the CreateClient.java can create a main client. Running inside the CreateServer.java can create a client, it is best to use jdk1.7 encoding format for UTF-8.